Addressing #128 moved the OWLOntologyChangeListener from InferredOWLClassHierarchyProvider to OWLReasonerManagerImpl and issuing fireReclassified() in case of logical axiom changes plus import changes (which were ignored before). This way all inferences in the UI should be updated, not only the inferred class hierarchy. Not sure the code was/is safe to guarantee that the reasoner processes changes before fireReclassified() because OWLManager processes listeners in undefined order. I do not think calling this method within SwingUtilities.infokeLater() makes it safe.
